+++ src/main/kotlin/com/christopherosthues/starwarsprogressbar/ui/events/VehicleClickListener.kt
@@ -1,0 +1,7 @@
+package com.christopherosthues.starwarsprogressbar.ui.events
+
+import com.christopherosthues.starwarsprogressbar.models.StarWarsVehicle
+
+internal interface VehicleClickListener {
+    fun vehicleClicked(vehicle: StarWarsVehicle)
+}